Make labels required.
