+++ assets/src/edit-story/components/form/color/color.js
@@ -68,7 +68,7 @@ ColorInput.propTypes = {
   hasGradient: PropTypes.bool,
   hasOpacity: PropTypes.bool,
   onChange: PropTypes.func.isRequired,
-  label: PropTypes.string,
+  label: PropTypes.string.isRequired,
   colorPickerActions: PropTypes.func,
 };
 
@@ -77,7 +77,6 @@ ColorInput.defaultProps = {
   hasGradient: false,
   hasOpacity: true,
   opacity: null,
-  label: null,
 };
 
 export default ColorInput;
