+++ assets/src/edit-story/components/form/dateTime/index.js
@@ -14,55 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies
- */
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import { rgba } from 'polished';
-import { useState } from 'react';
-
-/**
- * Internal dependencies
- */
-import TimePicker from './timePicker';
-import DatePicker from './datePicker';
-
-const DateTimeWrapper = styled.div`
-  box-shadow: 0 3px 30px rgba(25, 30, 35, 0.1);
-  border: 1px solid ${({ theme }) => rgba(theme.colors.bg.v0, 0.2)};
-  background-color: ${({ theme }) => theme.colors.fg.v1};
-  width: 100%;
-  padding: 4px;
-`;
-
-function DateTime({ value, onChange, is12Hour = true, forwardedRef }) {
-  const selectedTime = value ? new Date(value) : new Date();
-  const initialHours = selectedTime.getHours();
-  const [localData, setLocalData] = useState({
-    minutes: selectedTime.getMinutes(),
-    am: initialHours < 12 ? 'AM' : 'PM',
-    hours: is12Hour ? initialHours % 12 || 12 : initialHours,
-    date: selectedTime,
-  });
-  return (
-    <DateTimeWrapper ref={forwardedRef}>
-      <TimePicker
-        localData={localData}
-        setLocalData={setLocalData}
-        onChange={onChange}
-        is12Hour={is12Hour}
-      />
-      <DatePicker currentDate={value} onChange={onChange} />
-    </DateTimeWrapper>
-  );
-}
-
-DateTime.propTypes = {
-  onChange: PropTypes.func.isRequired,
-  value: PropTypes.string,
-  is12Hour: PropTypes.bool,
-  forwardedRef: PropTypes.object,
-};
-
-export default DateTime;
+export { default } from './dateTime';
