+++ assets/src/edit-story/components/form/numeric.js
@@ -178,7 +178,7 @@ Numeric.propTypes = {
   symbol: PropTypes.string,
   flexBasis: PropTypes.number,
   textCenter: PropTypes.bool,
-  ariaLabel: PropTypes.string,
+  ariaLabel: PropTypes.string.isRequired,
   float: PropTypes.bool,
 };
 
@@ -189,7 +189,6 @@ Numeric.defaultProps = {
   flexBasis: 110,
   textCenter: false,
   float: false,
-  ariaLabel: __('Standard input', 'web-stories'),
 };
 
 export default Numeric;
